# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/premium": {
      "get": {
        "operationId": "PremiumController_status",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Статус подписки",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PremiumStatusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "premium"
        ]
      }
    },
    "/v1/premium/subscribe": {
      "post": {
        "operationId": "PremiumController_subscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Подписка оформлена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PremiumStatusDto"
                }
              }
            }
          },
          "404": {
            "description": "Карта не найдена"
          },
          "409": {
            "description": "Подписка уже оформлена"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "premium"
        ]
      }
    },
    "/v1/premium/cancel": {
      "post": {
        "operationId": "PremiumController_cancel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Подписка отменена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PremiumStatusDto"
                }
              }
            }
          },
          "404": {
            "description": "Активной подписки нет"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "premium"
        ]
      }
    },
...
  },
  "info": {
    "title": "SmartQoldau API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PremiumStatusDto": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "description": "Есть ли доступ к Premium прямо сейчас. Остаётся true после отмены до конца оплаченного периода и во время ретраев автопродления (Р-09)."
          },
          "plan": {
            "type": "string",
            "enum": [
              "MONTH",
              "YEAR"
            ],
            "nullable": true
          },
          "currentPeriodEnd": {
            "type": "string",
            "nullable": true,
            "description": "Конец оплаченного периода"
          },
          "cancelled": {
            "type": "boolean",
            "description": "Отменена клиентом, не продлится"
          },
          "inGrace": {
            "type": "boolean",
            "description": "Оплата не прошла, идут ретраи — доступ пока сохраняется"
          }
        },
        "required": [
          "active",
          "plan",
          "currentPeriodEnd",
          "cancelled",
          "inGrace"
        ]
      },
      "SubscribeDto": {
        "type": "object",
        "properties": {
          "plan": {
            "type": "string",
            "enum": [
              "MONTH",
              "YEAR"
            ],
            "description": "Тариф Premium"
          },
          "paymentMethodId": {
            "type": "string",
            "description": "Привязанная карта, с которой списать"
          }
        },
        "required": [
          "plan",
          "paymentMethodId"
        ]
      },
...
    }
  }
}